move the loadAverage class into his own file and do the same for the specs
